Refactor of the Command component to extract the logic
As of now the Command creates a new application instance then looks
for all fixtures files, get the good loader for loading the fixtures
files and then get the Doctrine extractor depending of the database.

This is too much logic put on the Command component, it is very hard
to test it in a predictible way. As a result, this PR splits the
Command component into smaller bits, that may be reusable to
third-party libraries too.